+++ app/src/androidTest/java/cs7455/team5/dashyy/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package CS7455.Team5.dashyy;
+package cs7455.team5.dashyy;
 
 import android.content.Context;
