--- app/src/main/java/CS7455/Team5/dashyy/HomeActivity.java
@@ -1,54 +1,0 @@
-/*
-Mobile App Development - W01
-Team 5
-Summer 2020
-Semester Project - Dashyy
-
-Any use of the following code is forbidden without prior consent.
-*/
-
-package CS7455.Team5.dashyy;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.app.Activity;
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.ImageView;
-
-public class HomeActivity extends AppCompatActivity {
-
-    //Class Variables
-    ImageView maps_IV;
-
-    //Component Variables
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_home);
-
-        //Component Setup
-        maps_IV = findViewById(R.id.map_PB);
-        maps_IV.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(getApplicationContext(), MapsActivity.class);
-                startActivity(intent);
-            }
-        });
-    }
-
-    //SCREEN NAVIGATION
-
-    //Maps Screen
-    public void MapClick(View view)
-    {
-        Intent intent = new Intent (this, MapsActivity.class);
-        startActivity(intent);
-    }
-}
-
-
