--- app/src/main/java/CS7455/Team5/dashyy/SettingsActivity.java
@@ -1,36 +1,0 @@
-/*
-Mobile App Development - W01
-Team 5
-Summer 2020
-Semester Project - Dashyy
-
-Any use of the following code is forbidden without prior consent.
-*/
-
-package CS7455.Team5.dashyy;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-
-public class SettingsActivity extends AppCompatActivity {
-
-    //Class Variables
-
-
-    //Component Variables
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_settings);
-
-        //Component Setup
-    }
-
-    //SCREEN NAVIGATION
-
-}
