+++ app/src/main/java/cs7455/team5/dashyy/HomeActivity.java
@@ -1,0 +1,70 @@
+/*
+Mobile App Development - W01
+Team 5
+Summer 2020
+Semester Project - Dashyy
+
+Any use of the following code is forbidden without prior consent.
+*/
+
+package cs7455.team5.dashyy;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.MenuItem;
+import android.view.View;
+
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+
+public class HomeActivity extends AppCompatActivity {
+
+    //Class Variables
+
+
+    //Component Variables
+
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_home);
+
+        //Component Setup
+
+
+        //SCREEN NAVIGATION
+        BottomNavigationView bottomNavigationView = findViewById(R.id.bottom_navigation);
+        bottomNavigationView.setOnNavigationItemSelectedListener(new BottomNavigationView.OnNavigationItemSelectedListener() {
+            @Override
+            public boolean onNavigationItemSelected(@NonNull MenuItem item) {
+
+                switch (item.getItemId()) {
+                    case R.id.Maps_Page:
+                        Intent intent = new Intent(HomeActivity.this, MapsActivity.class);
+                        startActivity(intent);
+                        break;
+
+                    case R.id.Home_Page:
+                        break;
+
+                    case R.id.Settings_Page:
+                        Intent intent2 = new Intent(HomeActivity.this, SettingsActivity.class);
+                        startActivity(intent2);
+                        break;
+                }
+
+                return false;
+            }
+        });
+
+    }
+
+    public void MapsZoom (View view)
+    {
+        Intent intent = new Intent(this, MapsActivity.class);
+        startActivity(intent);
+    }
+}
