+++ app/src/main/java/cs7455/team5/dashyy/MapsActivity.java
@@ -1,4 +1,4 @@
-package CS7455.Team5.dashyy;
+package cs7455.team5.dashyy;
 
 import androidx.appcompat.app.AlertDialog;
 import androidx.fragment.app.FragmentActivity;
