+++ app/src/main/res/layout/activity_settings.xml
@@ -1,46 +1,23 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".SettingsActivity">
 
-    <ImageView
-        android:id="@+id/home_PB2"
-        android:layout_width="55dp"
-        android:layout_height="52dp"
-        android:layout_marginBottom="16dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:srcCompat="@drawable/ic_home_normal_24dp" />
-
-    <ImageView
-        android:id="@+id/map_PB2"
-        android:layout_width="55dp"
-        android:layout_height="52dp"
-        app:layout_constraintBottom_toBottomOf="@+id/home_PB2"
-        app:layout_constraintEnd_toStartOf="@+id/home_PB2"
-        app:layout_constraintStart_toStartOf="parent"
-        app:srcCompat="@drawable/ic_map_normal_24dp" />
-
-    <ImageView
-        android:id="@+id/settings_PB2"
-        android:layout_width="55dp"
-        android:layout_height="52dp"
-        app:layout_constraintBottom_toBottomOf="@+id/home_PB2"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/home_PB2"
-        app:srcCompat="@drawable/ic_settings_normal_24dp" />
-
-    <TextView
-        android:id="@+id/textView4"
-        android:layout_width="wrap_content"
+    <LinearLayout
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="48dp"
-        android:text="Settings Activity"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-</androidx.constraintlayout.widget.ConstraintLayout>
+        android:layout_gravity="bottom|center_horizontal"
+        android:orientation="vertical">
+
+        <com.google.android.material.bottomnavigation.BottomNavigationView
+            android:id="@+id/bottom_navigation"
+            style="@style/Widget.MaterialComponents.BottomNavigationView.Colored"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            app:menu="@menu/bottom_navigation_menu" />
+    </LinearLayout>
+
+</FrameLayout>
