+++ app/src/main/res/layout/activity_startup.xml
@@ -56,7 +56,7 @@
             android:layout_height="match_parent"
             android:layout_marginBottom="8dp"
             android:hint="Password"
-            app:endIconMode="password_toggle">>
+            app:endIconMode="password_toggle">
 
             <EditText
                 android:id="@+id/password_ET"
@@ -73,7 +73,7 @@
             android:layout_height="match_parent"
             android:hint="Confirm Password"
             android:visibility="invisible"
-            app:endIconMode="password_toggle">>
+            app:endIconMode="password_toggle">
 
             <EditText
                 android:id="@+id/passwordConfirm_ET"
