+++ app/src/test/java/cs7455/team5/dashyy/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package CS7455.Team5.dashyy;
+package cs7455.team5.dashyy;
 
 import org.junit.Test;
